Refactor ChunkManagerSource into ChunkStore + ChunkStoreView
