+++ packages/core/examples/chunk_streaming/chunk_info_overlay.ts
@@ -18,21 +18,24 @@ export class ChunkInfoOverlay {
 
   public update(idetik: Idetik, _timestamp?: DOMHighResTimeStamp): void {
     if (this.textDiv_.style.display === "none") return;
-    const chunkManagerSource = this.imageLayer_.chunkManagerSource;
-    if (!chunkManagerSource) {
-      this.textDiv_.textContent = "No chunk manager source";
+    const chunkStore = this.imageLayer_.chunkStore;
+    const chunkStoreView = this.imageLayer_.chunkStoreView;
+    if (!chunkStore || !chunkStoreView) {
+      this.textDiv_.textContent = "No chunk store";
       return;
     }
 
-    const chunksAtCurrentTime = chunkManagerSource.getChunksAtCurrentTime();
+    const chunksAtCurrentTime = chunkStore.getChunksAtTime(
+      chunkStore.getTimeIndex(this.imageLayer_.sliceCoords)
+    );
     if (!chunksAtCurrentTime) {
       this.textDiv_.textContent = "No chunks available";
       return;
     }
 
     const chunkDetails: string[] = [];
-    const currentLOD = chunkManagerSource.currentLOD;
-    const renderedChunks = chunkManagerSource.getChunks();
+    const currentLOD = chunkStoreView.currentLOD;
+    const renderedChunks = chunkStoreView.getChunks(this.imageLayer_.sliceCoords);
     const totalChunks = chunksAtCurrentTime.length;
 
     let loadedChunks = 0;
@@ -41,7 +44,7 @@ export class ChunkInfoOverlay {
       visible: number;
       rendered: number;
       prefetched: number;
-    }[] = Array.from({ length: chunkManagerSource.lodCount }, () => ({
+    }[] = Array.from({ length: chunkStore.lodCount }, () => ({
       visible: 0,
       rendered: 0,
       prefetched: 0,
