+++ packages/core/src/core/layer.ts
@@ -3,6 +3,7 @@ import { RenderableObject } from "./renderable_object";
 import { clamp } from "../utilities/clamp";
 import { Logger } from "../utilities/logger";
 import { EventContext } from "./event_dispatcher";
+import { Viewport } from "./viewport";
 
 export type LayerState = "initialized" | "loading" | "ready";
 export type blendMode = "normal" | "additive" | "subtractive" | "multiply";
@@ -17,6 +18,10 @@ export interface LayerOptions {
   opacity?: number;
   blendMode?: blendMode;
 }
+
+export type RenderContext = {
+  viewport: Viewport;
+};
 
 export abstract class Layer {
   public abstract readonly type: string;
@@ -59,7 +64,7 @@ export abstract class Layer {
     this.opacity_ = clamp(value, 0.0, 1.0);
   }
 
-  public abstract update(): void;
+  public abstract update(context?: RenderContext): void;
 
   public onEvent(_: EventContext): void {}
