+++ packages/core/src/data/chunk.ts
@@ -27,14 +27,17 @@ export function isChunkData(value: unknown): value is ChunkData {
   return false;
 }
 
-export type Chunk = {
-  data?: ChunkData;
-  state: "unloaded" | "queued" | "loading" | "loaded";
-  lod: number;
+export type ChunkViewState = {
   visible: boolean;
   prefetch: boolean;
   priority: number | null;
   orderKey: number | null;
+};
+
+export type Chunk = {
+  data?: ChunkData;
+  state: "unloaded" | "queued" | "loading" | "loaded";
+  lod: number;
   shape: {
     x: number;
     y: number;
@@ -60,7 +63,7 @@ export type Chunk = {
     y: number;
     z: number;
   };
-};
+} & ChunkViewState;
 
 // Maps Idetik spatial dimensions (x, y, z) and non-spatial dimensions (c, t)
 // dimensions to a chunk source's dimensions.
