+++ packages/core/src/renderers/webgl_renderer.ts
@@ -83,7 +83,7 @@ export class WebGLRenderer extends Renderer {
 
     this.state_.setDepthMask(true);
     for (const layer of opaque) {
-      layer.update();
+      layer.update({ viewport });
       if (layer.state === "ready") {
         this.renderLayer(layer, viewport.camera);
       }
@@ -91,7 +91,7 @@ export class WebGLRenderer extends Renderer {
 
     this.state_.setDepthMask(false);
     for (const layer of transparent) {
-      layer.update();
+      layer.update({ viewport });
       if (layer.state !== "ready") continue;
       this.renderLayer(layer, viewport.camera);
     }
