+++ packages/core/src/utilities/logger.ts
@@ -21,6 +21,8 @@ type Module =
   | "ChunkManager"
   | "ChunkManagerSource"
   | "ChunkQueue"
+  | "ChunkStore"
+  | "ChunkStoreView"
   | "ChunkedImageLayer"
   | "EventDispatcher"
   | "Idetik"
